backend/pyproject.toml
[project]
name = "lostfound-backend"
version = "0.1.0"
description = "AI-powered university lost and found backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.17.1",
    "email-validator>=2.3.0",
    "fastapi>=0.121.1",
    "httpx>=0.28.1",
    "numpy>=2.3.4",
    "passlib[bcrypt]>=1.7.4",
    "pgvector>=0.4.1",
    "pillow>=12.0.0",
    "psycopg[binary,pool]>=3.2.12",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "rapidfuzz>=3.10.1",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.44",
    "structlog>=25.5.0",
    "torch>=2.9.1",
    "transformers>=4.57.1",
    "uvicorn[standard]>=0.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

